add product controller in front end
